[refactor/TK-187]: RoleType을 domain으로 이동
