+++ src/main/java/withbeetravel/domain/User.java
@@ -7,7 +7,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.springframework.security.crypto.bcrypt.BCrypt;
-import withbeetravel.dto.request.auth.RoleType;
 
 @Entity
 @Getter
